requirements.txt
Flask==3.1.2
llama_parse==0.6.66
python-dotenv==1.1.1
